Server set up in server.js
First of I installed the required packages via the terminal (Express, Body-Parser and Cors) and included them in the server.js file.

I then set up the server with the port and listening() callback function at the bottom of server.js.

I then made the server the .listen() method and passed in the port number and listening callback function as parameters.

I got feedback in the terminal after calling 'node server.js', that the server is running correctly.