getData function (api call)
Here I've written the getData function which makes a call to the openweathermap api.

This is an async function that waits on the result of the api request. If the data is successfully retrieved, it is logged to the console and returnedin the try() area. If not the error is caught in the catch() area.

This function is triggered by an click event on the generate button. It is working as of this commit, with weather data successfully being logged to the console from the api.

In the next commit I'll continue to work on the chained promise which will store this data in the web app.