postData async function
In this commit I've written the async function postData which is chained to be called after the getData() function via .then().

This function has two parameters, the path and the dataObject. The path is the post route in the server.js file. The dataObject consists of three key:value pairs, the temperature (retrieved from the api call), the date (from the newDate variable in the app.js file), and the userResponse (the input from the html with an id of 'feelings' .)

Via the post route in the server.js file (addData), this data is turned into a newEntry object, and is then pushed to the projectData array.

To make sure everything was getting stored in the projectData array, I called the function several time, each time logging projectData to the console after the newItem had been stored. It successfully logged the previous entries to console showing that they are being stored in projectData.

Now that the data is being stored successfully, in the next commit I'll work on using this data to dynamically update the ui to the user.